Rafactor processTelegramWebhook
